reworked click to open add project hours modal, need to refactor modal into a timsheet-day component
