+++ src/app/app.module.ts
@@ -20,8 +20,7 @@ import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
 import { LoginComponent } from './login/login.component';
 import { NavbarComponent } from './navbar/navbar.component';
 import { TimesheetAddEditComponent } from './timesheet-add-edit/timesheet-add-edit.component';
-import { TimesheetDayComponent } from './Timesheet-Add-Edit/timesheet-day/timesheet-day.component';
-import { ProjectHoursListComponent } from './Timesheet-Add-Edit/timesheet-day/project-hours-list/project-hours-list.component';
+import { TimesheetDayComponent } from './timesheet-day/timesheet-day.component';
 import { TimesheetHistoryComponent } from './timesheet-history/timesheet-history.component';
 import { TimesheetHistoryProjectComponent } from './timesheet-history/timesheet-history-project/timesheet-history-project.component';
 import { TimesheetMainComponent } from './timesheet-main/timesheet-main.component';
@@ -40,7 +39,6 @@ import { FlatpickrModule } from 'angularx-flatpickr';
     NavbarComponent,
     TimesheetAddEditComponent,
     TimesheetDayComponent,
-    ProjectHoursListComponent,
     TimesheetHistoryComponent,
     TimesheetHistoryProjectComponent,
     TimesheetMainComponent
