+++ src/app/timesheet-add-edit/timesheet-add-edit.component.html
@@ -17,19 +17,6 @@
   <div class="col-md-4">
     <h3>{{ viewDate | calendarDate:(view + 'ViewTitle'):'en' }}</h3>
   </div>
-  <!-- <div class="col-md-4">
-    <div class="btn-group">
-      <div class="btn btn-primary" (click)="setView(CalendarView.Month)" [class.active]="view === CalendarView.Month">
-        Month
-      </div>
-      <div class="btn btn-primary" (click)="setView(CalendarView.Week)" [class.active]="view === CalendarView.Week">
-        Week
-      </div>
-      <div class="btn btn-primary" (click)="setView(CalendarView.Day)" [class.active]="view === CalendarView.Day">
-        Day
-      </div>
-    </div>
-  </div> -->
 </div>
 
 
@@ -39,104 +26,60 @@
     [refresh]="refresh" [activeDayIsOpen]="activeDayIsOpen" (dayClicked)="dayClicked($event.day)"
     (eventClicked)="handleEvent('Clicked', $event.event)" (eventTimesChanged)="eventTimesChanged($event)">
   </mwl-calendar-month-view>
-  <mwl-calendar-week-view *ngSwitchCase="CalendarView.Week" [viewDate]="viewDate" [events]="events" [refresh]="refresh"
-    (eventClicked)="handleEvent('Clicked', $event.event)" (eventTimesChanged)="eventTimesChanged($event)">
-  </mwl-calendar-week-view>
-  <mwl-calendar-day-view *ngSwitchCase="CalendarView.Day" [viewDate]="viewDate" [events]="events" [refresh]="refresh"
-    (eventClicked)="handleEvent('Clicked', $event.event)" (eventTimesChanged)="eventTimesChanged($event)">
-  </mwl-calendar-day-view>
 </div>
 
 
 <!-- if day clicked, modal show app-timesheet-day component -->
-<!-- <app-timesheet-day></app-timesheet-day> -->
 <!-- Everything you see below is just for the demo, you don't need to include it in your app -->
-
-<br /><br /><br />
-
-<h3>
-  Add Project Hours
-  <button class="btn btn-primary float-right ml-4">
-    <!-- <button class="btn btn-primary float-right" (click)="submitHours()"> CREATE NEW METHODS TO INTERACT WITH DATABASE-->
-    Submit
-  </button>
-  <button class="btn btn-primary float-right ">
-    <!-- <button class="btn btn-primary float-right" (click)="saveHours()"> CREATE NEW METHODS TO INTERACT WITH DATABASE-->
-    Save
-  </button>
-
-  <div class="clearfix"></div>
-</h3>
-
-
-<div class="table-responsive">
-  <table class="table table-bordered">
-    <thead>
-      <tr>
-        <th>Project</th>
-        <th>Hours</th>
-        <!-- <th>Secondary color</th>
-        <th>Starts at</th>
-        <th>Ends at</th>
-        <th>Remove</th> -->
-      </tr>
-    </thead>
-
-    <tbody>
-      <tr *ngFor="let event of events">
-        <td>
-          <input type="text" class="form-control" [(ngModel)]="event.title" (keyup)="refresh.next()" />
-          <!-- DROP DOWN LIST OF PROJECTS FORM THE DATABASE -->
-        </td>
-        <!-- <td>
-          <input type="color" [(ngModel)]="event.color.primary" (change)="refresh.next()" />
-        </td>
-        <td>
-          <input type="color" [(ngModel)]="event.color.secondary" (change)="refresh.next()" />
-        </td>
-        <td>
-          <input class="form-control" type="text" mwlFlatpickr [(ngModel)]="event.start"
-            (ngModelChange)="refresh.next()" [altInput]="true" [convertModelValue]="true" [enableTime]="true"
-            dateFormat="Y-m-dTH:i" altFormat="F j, Y H:i" placeholder="Not set" />
-        </td> -->
-        <!-- <td>
-          <input class="form-control" type="text" mwlFlatpickr [(ngModel)]="event.end" (ngModelChange)="refresh.next()"
-            [altInput]="true" [convertModelValue]="true" [enableTime]="true" dateFormat="Y-m-dTH:i"
-            altFormat="F j, Y H:i" placeholder="Not set" />
-        </td> -->
-        <td>
-          <input class="form-control" type="number" />
-        </td>
-        <!-- <td>
-          <button class="btn btn-danger" (click)="deleteEvent(event)">
-            Delete
-          </button>
-        </td> -->
-      </tr>
-    </tbody>
-  </table>
-</div>
 
 <ng-template #modalContent let-close="close">
   <div class="modal-header">
-    <h5 class="modal-title">Event action occurred</h5>
+    <h5 class="modal-title">Add Project Hours</h5>
     <button type="button" class="close" (click)="close()">
       <span aria-hidden="true">&times;</span>
     </button>
   </div>
   <div class="modal-body">
-    <div>
-      Action:
-      <pre>{{ modalData?.action }}</pre>
+    <div class="table-responsive">
+      <table class="table table-bordered">
+        <thead>
+          <tr>
+            <th>Project</th>
+            <th>Hours</th>
+
+          </tr>
+        </thead>
+        <tbody>
+          <tr *ngFor="let event of events">
+            <td>
+              <input type="text" class="form-control" [(ngModel)]="event.title" (keyup)="refresh.next()" />
+              <!-- DROP DOWN LIST OF PROJECTS FORM THE DATABASE -->
+            </td>
+
+            <td>
+              <input class="form-control" type="number" />
+            </td>
+          </tr>
+        </tbody>
+      </table>
     </div>
-    <div>
-      Event:
-      <pre>{{ modalData?.event | json }}</pre>
-    </div>
+
+    <h3 class="">Total Hours for DATE</h3>
+
+    <h4 class="float-right">45 Hours</h4>
+
   </div>
+
   <div class="modal-footer">
+    <!-- (click)="submitHours()" CREATE NEW METHODS TO INTERACT WITH DATABASE-->
+    <button class="btn btn-primary float-right ml-4">
+      Submit </button>
+    <!-- (click)="saveHours()" CREATE NEW METHODS TO INTERACT WITH DATABASE-->
+    <button class="btn btn-primary float-right ">
+      Save </button>
     <button type="button" class="btn btn-outline-secondary" (click)="close()">
-      OK
+      Close
     </button>
   </div>
+
 </ng-template>
