--- src/app/timesheet-add-edit/timesheet-day/project-hours-list/project-hours-list.component.html
@@ -1,1 +1,0 @@
-<p>project-hours-list works!</p>
