--- src/app/timesheet-add-edit/timesheet-day/project-hours-list/project-hours-list.component.ts
@@ -1,15 +1,0 @@
-import { Component, OnInit } from '@angular/core';
-
-@Component({
-  selector: 'app-project-hours-list',
-  templateUrl: './project-hours-list.component.html',
-  styleUrls: ['./project-hours-list.component.scss']
-})
-export class ProjectHoursListComponent implements OnInit {
-
-  constructor() { }
-
-  ngOnInit(): void {
-  }
-
-}
