--- src/app/timesheet-add-edit/timesheet-day/timesheet-day.component.ts
@@ -1,37 +1,0 @@
-import { Component, OnInit } from '@angular/core';
-import { NgbModal, ModalDismissReasons } from '@ng-bootstrap/ng-bootstrap';
-
-@Component({
-  selector: 'app-timesheet-day',
-  templateUrl: './timesheet-day.component.html',
-  styleUrls: ['./timesheet-day.component.scss']
-})
-export class TimesheetDayComponent implements OnInit {
-  closeResult = '';
-
-  constructor(private modalService: NgbModal) { }
-
-  open(content) {
-    this.modalService.open(content, { ariaLabelledBy: 'modal-basic-title' }).result.then((result) => {
-      this.closeResult = `Closed with: ${result}`;
-    }, (reason) => {
-      this.closeResult = `Dismissed ${this.getDismissReason(reason)}`;
-    });
-  }
-
-  ngOnInit(): void {
-  }
-
-  private getDismissReason(reason: any): string {
-    if (reason === ModalDismissReasons.ESC) {
-      return 'by pressing ESC';
-    } else if (reason === ModalDismissReasons.BACKDROP_CLICK) {
-      return 'by clicking on a backdrop';
-    } else {
-      return `with: ${reason}`;
-    }
-  }
-
-  // put/post Requests go in here
-
-}
