+++ src/app/timesheet-day/timesheet-day.component.ts
@@ -1,0 +1,18 @@
+import { Component, OnInit } from '@angular/core';
+
+
+@Component({
+  selector: 'app-timesheet-day',
+  templateUrl: './timesheet-day.component.html',
+  styleUrls: ['./timesheet-day.component.scss']
+})
+export class TimesheetDayComponent implements OnInit {
+
+
+  constructor() { }
+
+
+  ngOnInit(): void {
+  }
+
+}
